+++ 2015/src/common/utils/utils.h
@@ -1,9 +1,11 @@
 #pragma once
 
+#include <stdint.h>
+
 char* file_to_string(const char*);
 
 void   free_array(void**);
 char** file_to_array(const char*);
 
-int min(int, ...);
-int max(int, ...);
+int64_t min(uint8_t, ...);
+int64_t max(uint8_t, ...);
